Explicit/persistent controller numbers

+++ macOS/ControllerFilePipe.swift
@@ -6,7 +6,7 @@ private var newline: UInt8 = 0x0A
 final class ControllerFilePipe {
     private var outputStream: OutputStream
     
-    init(index: Int) throws {
+    init(index: UInt8) throws {
         self.outputStream = try createPipe(index: index)
         DispatchQueue.global().async {
             self.outputStream.open()
